fix: solve previous error in insertDefaults.sql
The SQL code block was interpreted as declared incorrectly because the DECLARE section was not properly terminated with '$$' after the first ';' character. To avoid any interpretation issues by jdbc, the block was declared using single quotes ('), making it so that the content inside the block is not interpreted and therefore causing no errors.